requirements.txt
requests
beautifulsoup4
yfinance
pyyaml
playwright
